+++ src/wmclient.cc
@@ -1951,8 +1951,8 @@ bool YFrameClient::getNetWMUserTime(Window window, unsigned long &time) {
 
             MSG(("got user time"));
             time = utime[0] & 0xffffffff;
-            if (time == -1UL)
-                    time = -2UL;
+            if (time == (unsigned long) -1)
+                    time = (unsigned long) -2;
 
             XFree(prop);
             return true;
